phase-2-web/backend/requirements.txt
# Hackathon II - Phase 2 Backend Requirements
# Production-ready with authentication and database

# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.5.2

# Database & ORM
sqlmodel>=0.0.22
asyncpg>=0.30.0
alembic>=1.13.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# Testing (Optional)
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
